make a single orchestrator class. bug fixes found in result.
