+++ chat_toolkit/__init__.py
@@ -1,6 +1,5 @@
-from .common import set_openai_api_key
-from .components import OpenAIChatBot, OpenAISpeechToText
-from .orchestrators import SpeechToTextOrchestrator, TextToTextOrchestrator
+from .common import Orchestrator, set_openai_api_key
+from .components import OpenAIChatBot, OpenAISpeechToText, Pyttsx3TextToSpeech
 
 __version__ = "1.0.1"
 
@@ -8,6 +7,6 @@ __all__ = (
     "set_openai_api_key",
     "OpenAIChatBot",
     "OpenAISpeechToText",
-    "SpeechToTextOrchestrator",
-    "TextToTextOrchestrator",
+    "Orchestrator",
+    "Pyttsx3TextToSpeech",
 )
