+++ chat_toolkit/common/__init__.py
@@ -1,10 +1,12 @@
 from .constants import TMP_DIR
 from .custom_types import StartingPromptsType
+from .orchestrator import Orchestrator
 from .utils import set_openai_api_key, temporary_file
 
 __all__ = (
     "set_openai_api_key",
     "temporary_file",
+    "Orchestrator",
     "StartingPromptsType",
     "TMP_DIR",
 )
